Implement inlining of function calls
Function calls are inlined to perform type inference in the context of the
caller.

The commit also contains the change that equals does not need to be true
even if the types have the same id. The reason for this is, that
a modified type is never equal to the other with the same id. The equals
funciton is used to calculate the difference of two type environments and
if types with the same id are equal, then changes to these are not
shown as differences.

Modification to a record type (like adding or changing properties)
returns now a record type that has the same id as the original one.
This is needed to be able to detect if an argument type in a function
call has changed or not. If the id would change, then there is no way
to detect if the type has been modified (e.g. adding properties) or a
complete new type has been assigned.

fixes #34