Implement a more sophistic recursion detection
This recursion detection tries to figure out if a previous call
used exactly the same arguments. If this is the case, then it can
be expected that both calls will have the same return type. So instead
of evaluating the function again, use the same type variable as return
type as the previous call does. If the previous call completes, then
the return type is substituted.